"reorganized folders and files. added some notes in some folders. some folders in patient and admin needs be renamed such as "section" and "containers". will need to reroute admin and update its sidebar."
