--- client/src/components/dashboards/PatientDashboard.txt
@@ -1,21 +1,0 @@
-import React from "react";
-
-const PatientDashboardOld = () => {
-  return (
-    <div>
-      <h1>Patient Dashboard</h1>
-      <p>
-        Welcome, Patient! Here you can view your appointments, medical history,
-        and prescriptions.
-      </p>
-      <ul>
-        <li>View Appointments</li>
-        <li>Check Medical History</li>
-        <li>View Prescriptions</li>
-        <li>Update Profile</li>
-      </ul>
-    </div>
-  );
-};
-
-export default PatientDashboardOld;
