+++ client/src/components/users/admin/sections/doctorRegistration.jsx
