+++ client/src/components/users/patient/sections/header/IconLogoPatient.jsx
