+++ client/src/components/users/patient/sidebar/pages/medical-history/sections/Allergies.tsx
