+++ client/src/components/users/patient/sidebar/pages/medical-history/sections/Surgeries.tsx
