+++ client/src/routers/AdminRoutes.jsx
@@ -1,11 +1,11 @@
 import React from "react";
 import { Routes, Route } from "react-router-dom";
-import AdminDashboard from "../components/dashboards/admin/AdminDashboard";
-import RegisterDoctor from "../components/dashboards/admin/sections/doctorRegistration";
-import ManageDoctors from "../components/dashboards/admin/sections/manageDoctors";
-import ManagePatients from "../components/dashboards/admin/sections/managePatients";
-import Settings from "../components/dashboards/admin/sections/settings";
-import SystemReports from "../components/dashboards/admin/sections/system-reports";
+import AdminDashboard from "../components/users/admin/AdminDashboard";
+import RegisterDoctor from "../components/users/admin/sections/doctorRegistration";
+import ManageDoctors from "../components/users/admin/sections/manageDoctors";
+import ManagePatients from "../components/users/admin/sections/managePatients";
+import Settings from "../components/users/admin/sections/settings";
+import SystemReports from "../components/users/admin/sections/system-reports";
 
 const AdminRoutes = () => {
   return (
